Split widget: computed header
